requirements.txt
django==1.4
Mezzanine==1.2.2
psycopg2==2.4.5
py-bcrypt==0.2
python-memcached==1.48
south==0.7.4

# Debug Tools
sqlparse==0.1.3
django-debug-toolbar==0.9.4
django-devserver==0.3.1

# Deployment
fabric==1.4.1

# AWS
boto==2.3.0
django-storages==1.1.4
django-ses==0.4.1
django_smtp_ssl

# ECL
https://packages.elmcitylabs.com/ecl_django-0.4.16.tar.gz#ecl_django
https://packages.elmcitylabs.com/ecl_tools-0.3.7.tar.gz#ecl_tools
# https://packages.elmcitylabs.com/chargemaster-0.2.19.tar.gz
# ecl-facebook==1.2.7
# ecl-twitter==1.2.0

# Search
# django-haystack==1.2.6
# pysolr==2.1.0-beta
# whoosh==2.3.2

# Misc
PIL
# django-shorturls==1.0.1
# suds==0.4
# django-celery==2.5.1

django-redis-sessions==0.2.3

sorl-thumbnail==11.12

django-registration

#caching
hiredis
redis
django-redis-cache
#doesn't seem to like grappelli
django-redis-status

django-durationfield
django-avatar

ipython
django-shortcuts
django-extensions
